fix(powerforge): 🐛 Preserve relocated MSI imports
