+++ PowerForge/Services/DotNetPublishPipelineRunner.MsiBuildWorkspace.BuildConfiguration.cs
@@ -36,6 +36,7 @@ public sealed partial class DotNetPublishPipelineRunner
         var plannedSources = new List<string>();
         var nuGetConfigs = new List<string>();
         var visibleConfigNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+        var projectRootDirectory = AppendDirectorySeparator(Path.GetFullPath(projectRootPath));
         var queue = new Queue<string>();
         foreach (var directory in GetGeneratedInstallerBuildConfigurationDirectories(projectRootPath, sourceProjectPath))
         {
@@ -52,6 +53,7 @@ public sealed partial class DotNetPublishPipelineRunner
                 var targetPath = GetGeneratedInstallerVisibleBuildConfigurationTargetPath(
                     file,
                     workingDirectory,
+                    projectRootDirectory,
                     visibleConfigNames);
                 if (!plannedTargets.ContainsKey(sourceFullPath))
                 {
@@ -81,6 +83,7 @@ public sealed partial class DotNetPublishPipelineRunner
     private static string GetGeneratedInstallerVisibleBuildConfigurationTargetPath(
         string sourcePath,
         string workingDirectory,
+        string projectRootDirectory,
         ISet<string> visibleConfigNames)
     {
         var fileName = Path.GetFileName(sourcePath);
@@ -89,11 +92,13 @@ public sealed partial class DotNetPublishPipelineRunner
             return Path.Combine(workingDirectory, fileName);
         }
 
+        var relativePath = Path.GetFullPath(sourcePath)
+            .Substring(AppendDirectorySeparator(Path.GetFullPath(projectRootDirectory)).Length);
         return Path.Combine(
             workingDirectory,
             "PowerForgeInputs",
             "BuildConfig",
-            fileName);
+            relativePath);
     }
 
     private static IEnumerable<string> GetGeneratedInstallerBuildConfigurationDirectories(
@@ -145,7 +150,7 @@ public sealed partial class DotNetPublishPipelineRunner
         copiedFiles[sourceFullPath] = targetPath;
         RebaseNuGetConfigPaths(targetPath, Path.GetDirectoryName(sourceFullPath)!);
         RewriteGeneratedInstallerBuildFileDirectoryProperties(targetPath, sourceFullPath);
-        RewriteGeneratedInstallerBuildPropertyFunctionImports(targetPath, sourceFullPath, projectRootDirectory, workingDirectory, copiedFiles, plannedTargets);
+        RewriteGeneratedInstallerBuildImports(targetPath, sourceFullPath, projectRootDirectory, workingDirectory, copiedFiles, plannedTargets);
         queue.Enqueue(sourceFullPath);
     }
 
@@ -334,306 +339,6 @@ public sealed partial class DotNetPublishPipelineRunner
              char.IsLetter(value[0]) &&
              value[1] == ':' &&
              (value[2] == '\\' || value[2] == '/'));
-    }
-
-    private static IEnumerable<string> ResolveGeneratedInstallerBuildImports(string sourcePath, string projectRoot)
-    {
-        XDocument document;
-        try
-        {
-            document = XDocument.Load(sourcePath, LoadOptions.PreserveWhitespace);
-        }
-        catch
-        {
-            yield break;
-        }
-
-        var sourceDirectory = Path.GetDirectoryName(sourcePath)!;
-        var projectRootDirectory = AppendDirectorySeparator(Path.GetFullPath(projectRoot));
-        foreach (var import in document
-            .Descendants()
-            .Where(element => string.Equals(element.Name.LocalName, "Import", StringComparison.OrdinalIgnoreCase))
-            .Select(element => element.Attribute("Project"))
-            .Where(attribute => attribute is not null)
-            .Select(attribute => attribute!.Value)
-            .Where(value => !string.IsNullOrWhiteSpace(value)))
-        {
-            foreach (var path in ResolveGeneratedInstallerBuildImport(import, sourceDirectory, projectRootDirectory))
-            {
-                yield return path;
-            }
-        }
-    }
-
-    private static IEnumerable<string> ResolveGeneratedInstallerBuildImport(
-        string import,
-        string sourceDirectory,
-        string projectRootDirectory)
-    {
-        var expanded = NormalizeRelativePathSeparators(ExpandGeneratedInstallerBuildImport(import, sourceDirectory, projectRootDirectory));
-        foreach (var path in ResolveGeneratedInstallerGetPathOfFileAboveImport(expanded, sourceDirectory, projectRootDirectory))
-        {
-            yield return path;
-        }
-
-        if (expanded.IndexOf("$(", StringComparison.Ordinal) >= 0 ||
-            expanded.IndexOf("%(", StringComparison.Ordinal) >= 0)
-        {
-            yield break;
-        }
-
-        var importPath = Path.IsPathRooted(expanded)
-            ? expanded
-            : Path.Combine(sourceDirectory, expanded);
-        var projectRoot = AppendDirectorySeparator(Path.GetFullPath(projectRootDirectory));
-        if (importPath.IndexOfAny(new[] { '*', '?' }) >= 0)
-        {
-            var searchOption = SearchOption.TopDirectoryOnly;
-            var recursiveIndex = importPath.IndexOf("**", StringComparison.Ordinal);
-            var directory = Path.GetDirectoryName(importPath);
-            var pattern = Path.GetFileName(importPath);
-            if (recursiveIndex >= 0)
-            {
-                var prefix = importPath.Substring(0, recursiveIndex).TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
-                directory = string.IsNullOrWhiteSpace(prefix)
-                    ? sourceDirectory
-                    : prefix;
-                searchOption = SearchOption.AllDirectories;
-            }
-            if (string.IsNullOrWhiteSpace(directory) ||
-                !Directory.Exists(directory))
-            {
-                yield break;
-            }
-
-            foreach (var file in Directory.EnumerateFiles(directory, pattern, searchOption))
-            {
-                var fullPath = Path.GetFullPath(file);
-                if (fullPath.StartsWith(projectRoot, StringComparison.OrdinalIgnoreCase))
-                {
-                    yield return fullPath;
-                }
-            }
-
-            yield break;
-        }
-
-        var fullImportPath = Path.GetFullPath(importPath);
-        if (fullImportPath.StartsWith(projectRoot, StringComparison.OrdinalIgnoreCase) &&
-            File.Exists(fullImportPath))
-        {
-            yield return fullImportPath;
-        }
-    }
-
-    private static IEnumerable<string> ResolveGeneratedInstallerGetPathOfFileAboveImport(
-        string import,
-        string sourceDirectory,
-        string projectRootDirectory)
-    {
-        const string marker = "[MSBuild]::GetPathOfFileAbove(";
-        var markerIndex = import.IndexOf(marker, StringComparison.OrdinalIgnoreCase);
-        if (markerIndex < 0)
-        {
-            yield break;
-        }
-
-        var startIndex = markerIndex + marker.Length;
-        var endIndex = FindGeneratedInstallerBuildFunctionEnd(import, startIndex);
-        if (endIndex < 0)
-        {
-            yield break;
-        }
-
-        var arguments = SplitGeneratedInstallerBuildFunctionArguments(import.Substring(startIndex, endIndex - startIndex));
-        if (arguments.Count == 0)
-        {
-            yield break;
-        }
-
-        var fileName = TrimGeneratedInstallerBuildFunctionArgument(arguments[0]);
-        if (string.IsNullOrWhiteSpace(fileName) ||
-            fileName.IndexOfAny(new[] { Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar }) >= 0)
-        {
-            yield break;
-        }
-
-        var searchDirectory = sourceDirectory;
-        if (arguments.Count > 1)
-        {
-            var startDirectory = TrimGeneratedInstallerBuildFunctionArgument(arguments[1]);
-            if (!string.IsNullOrWhiteSpace(startDirectory) &&
-                startDirectory.IndexOf("$(", StringComparison.Ordinal) < 0 &&
-                startDirectory.IndexOf("%(", StringComparison.Ordinal) < 0)
-            {
-                searchDirectory = Path.IsPathRooted(startDirectory)
-                    ? startDirectory
-                    : Path.Combine(sourceDirectory, startDirectory);
-            }
-        }
-
-        var projectRoot = AppendDirectorySeparator(Path.GetFullPath(projectRootDirectory));
-        var currentDirectory = Path.GetFullPath(searchDirectory);
-        while (AppendDirectorySeparator(currentDirectory).StartsWith(projectRoot, StringComparison.OrdinalIgnoreCase))
-        {
-            var candidate = Path.Combine(currentDirectory, fileName);
-            if (File.Exists(candidate))
-            {
-                yield return Path.GetFullPath(candidate);
-                yield break;
-            }
-
-            var parent = Path.GetDirectoryName(currentDirectory);
-            if (string.IsNullOrWhiteSpace(parent) ||
-                string.Equals(parent, currentDirectory, StringComparison.OrdinalIgnoreCase))
-            {
-                yield break;
-            }
-
-            currentDirectory = parent;
-        }
-    }
-
-    private static int FindGeneratedInstallerBuildFunctionEnd(string value, int startIndex)
-    {
-        var quote = '\0';
-        for (var index = startIndex; index < value.Length; index++)
-        {
-            var current = value[index];
-            if (quote != '\0')
-            {
-                if (current == quote)
-                {
-                    quote = '\0';
-                }
-
-                continue;
-            }
-
-            if (current == '\'' || current == '"')
-            {
-                quote = current;
-                continue;
-            }
-
-            if (current == ')')
-            {
-                return index;
-            }
-        }
-
-        return -1;
-    }
-
-    private static List<string> SplitGeneratedInstallerBuildFunctionArguments(string arguments)
-    {
-        var values = new List<string>();
-        var quote = '\0';
-        var startIndex = 0;
-        for (var index = 0; index < arguments.Length; index++)
-        {
-            var current = arguments[index];
-            if (quote != '\0')
-            {
-                if (current == quote)
-                {
-                    quote = '\0';
-                }
-
-                continue;
-            }
-
-            if (current == '\'' || current == '"')
-            {
-                quote = current;
-                continue;
-            }
-
-            if (current == ',')
-            {
-                values.Add(arguments.Substring(startIndex, index - startIndex));
-                startIndex = index + 1;
-            }
-        }
-
-        values.Add(arguments.Substring(startIndex));
-        return values;
-    }
-
-    private static string TrimGeneratedInstallerBuildFunctionArgument(string value)
-    {
-        var trimmed = value.Trim();
-        if (trimmed.Length >= 2 &&
-            ((trimmed[0] == '\'' && trimmed[trimmed.Length - 1] == '\'') ||
-             (trimmed[0] == '"' && trimmed[trimmed.Length - 1] == '"')))
-        {
-            return trimmed.Substring(1, trimmed.Length - 2);
-        }
-
-        return trimmed;
-    }
-
-    private static void RewriteGeneratedInstallerBuildPropertyFunctionImports(
-        string targetPath,
-        string sourcePath,
-        string projectRootDirectory,
-        string workingDirectory,
-        IDictionary<string, string> copiedFiles,
-        IReadOnlyDictionary<string, string> plannedTargets)
-    {
-        XDocument document;
-        try
-        {
-            document = XDocument.Load(targetPath, LoadOptions.PreserveWhitespace);
-        }
-        catch
-        {
-            return;
-        }
-
-        var sourceDirectory = Path.GetDirectoryName(sourcePath)!;
-        var targetDirectory = Path.GetDirectoryName(targetPath)!;
-        var changed = false;
-        foreach (var projectAttribute in document
-            .Descendants()
-            .Where(element => string.Equals(element.Name.LocalName, "Import", StringComparison.OrdinalIgnoreCase))
-            .Select(element => element.Attribute("Project"))
-            .Where(attribute => attribute is not null))
-        {
-            var project = projectAttribute!.Value;
-            if (project.IndexOf("[MSBuild]::GetPathOfFileAbove(", StringComparison.OrdinalIgnoreCase) < 0)
-            {
-                continue;
-            }
-
-            var resolvedPath = ResolveGeneratedInstallerBuildImport(project, sourceDirectory, projectRootDirectory).FirstOrDefault();
-            if (string.IsNullOrWhiteSpace(resolvedPath))
-            {
-                continue;
-            }
-
-            var resolvedFullPath = Path.GetFullPath(resolvedPath);
-            var resolvedTargetPath = copiedFiles.TryGetValue(resolvedFullPath, out var existingTarget)
-                ? existingTarget
-                : plannedTargets.TryGetValue(resolvedFullPath, out var plannedTarget)
-                    ? plannedTarget
-                : GetGeneratedInstallerBuildFileTargetPath(resolvedFullPath, projectRootDirectory, workingDirectory);
-            var rewrittenProject = GetRelativePathCompat(targetDirectory, resolvedTargetPath);
-            projectAttribute.Value = rewrittenProject;
-            var condition = projectAttribute.Parent?.Attribute("Condition");
-            if (condition is not null &&
-                condition.Value.IndexOf("[MSBuild]::GetPathOfFileAbove(", StringComparison.OrdinalIgnoreCase) >= 0)
-            {
-                condition.Value = "Exists('" + rewrittenProject + "')";
-            }
-
-            changed = true;
-        }
-
-        if (changed)
-        {
-            document.Save(targetPath);
-        }
     }
 
     private static void RewriteGeneratedInstallerBuildFileDirectoryProperties(
